Remove hypervolume.utils and .base as they are unnecessary
