+++ optuna/_hypervolume/__init__.py
@@ -1,12 +1,5 @@
-from optuna._hypervolume.base import BaseHypervolume
 from optuna._hypervolume.hssp import _solve_hssp
-from optuna._hypervolume.utils import _compute_2d
 from optuna._hypervolume.wfg import WFG
 
 
-__all__ = [
-    "BaseHypervolume",
-    "_compute_2d",
-    "_solve_hssp",
-    "WFG",
-]
+__all__ = ["_solve_hssp", "WFG"]
