+++ tests/hypervolume_tests/test_wfg.py
@@ -15,15 +15,16 @@ def _shuffle_and_filter_sols(
 
 
 @pytest.mark.parametrize("assume_pareto", (True, False))
-def test_wfg_2d(assume_pareto: bool) -> None:
+@pytest.mark.parametrize("n_sols", list(range(2, 30)))
+def test_wfg_2d(assume_pareto: bool, n_sols: int) -> None:
+    n = n_sols
     rng = np.random.RandomState(42)
-    for n in range(2, 30):
-        r = n * np.ones(2)
-        s = np.empty((2 * n + 1, 2), dtype=int)
-        s[:n] = np.stack([np.arange(n), np.arange(n)[::-1]], axis=-1)
-        s[n:] = np.stack([np.arange(n + 1), np.arange(n + 1)[::-1]], axis=-1)
-        s = _shuffle_and_filter_sols(s, assume_pareto, rng)
-        assert optuna._hypervolume.WFG().compute(s, r, assume_pareto) == n * n - n * (n - 1) // 2
+    r = n * np.ones(2)
+    s = np.empty((2 * n + 1, 2), dtype=int)
+    s[:n] = np.stack([np.arange(n), np.arange(n)[::-1]], axis=-1)
+    s[n:] = np.stack([np.arange(n + 1), np.arange(n + 1)[::-1]], axis=-1)
+    s = _shuffle_and_filter_sols(s, assume_pareto, rng)
+    assert optuna._hypervolume.WFG().compute(s, r, assume_pareto) == n * n - n * (n - 1) // 2
 
 
 @pytest.mark.parametrize("n_objs", list(range(2, 10)))
